pyproject.toml
[project]
name = "car-color-graph"
version = "0.1.0"
description = "LangGraph supervisor/subgraph demo – the car is blue."
requires-python = ">=3.10"
dependencies = [
    "langgraph==0.4.8",
    "langgraph-supervisor==0.0.27",
    "langchain-core==0.3.65",
    "langchain-openai==0.3.22",
    "python-dotenv==1.1.0",
    "langgraph-cli[inmem]==0.3.3",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
